Move logDuration / timeFunction / DurationResult into timing.py
Move the three timing helpers out of `utils.py` and into the existing
`timing` module, where `BoxCarTimer` already lives. They naturally
belong together: all three measure elapsed time of either a block of
code or a function call, and they have the same minimal-dependency
profile (stdlib only).

The classes/functions are moved verbatim, preserving the existing
behaviour, docstrings and signatures. The module's `__all__` is
updated to export the new public symbols.

All callers (aos, guiders, pipelineRunning, plotting/mosaicing,
timedServices) have been updated to import from `timing` instead of
`utils`. The `contextmanager`, `dataclass`, `wraps`, `perf_counter`
and `Iterator` imports that those helpers needed have moved with
them; `utils.py` no longer needs any of them.

Verified:
- mypy and flake8 clean across the package, scripts, and tests.
- The full unit test suite (195 tests) passes — the same two
  pre-existing sleep-timing flakes (`test_max`, `test_median`,
  occasionally `test_min` / `test_mean`) skip cleanly and are
  unrelated to this change.
- Every dependent in-package module imports cleanly.

Written by Claude Code 4.6 Opus (high-effort).
